fix(tts): trim FluidAudio cache diagnostics
